evaluator,fitsinto: move fitsinto into evaluator due to cycle
